requirements.txt
python-telegram-bot
openai
gspread
oauth2client
requests
